requirements.txt
# Core dependencies
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-openai==0.1.8
langgraph==0.0.69
streamlit==1.35.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
asyncpg==0.30.0

# Environment and configuration
python-dotenv==1.0.1
toml==0.10.2

# Data manipulation
pandas==2.2.2
numpy==1.26.4

# Logging and utilities
loguru==0.7.2

# LLM providers
openai==1.35.3
together==1.0.12
huggingface-hub==0.20.3
ollama==0.5.3

# Streamlit async support
nest-asyncio==1.6.0

# API and web frameworks
fastapi==0.111.0
uvicorn==0.30.1
langserve==0.0.54

# Testing
pytest==8.2.2
pytest-asyncio==0.23.6
pytest-cov==7.0.0
black==24.4.2
flake8==7.0.0

# Additional utilities
pydantic==2.7.4
requests==2.31.0
aiohttp==3.9.5
aiofiles==23.2.1
httpx==0.27.0
plotly==5.22.0
dataclasses-json==0.6.7
typing-extensions==4.12.2
